stack  pt , offsets

+++ milestone2/tests/test_1.java
@@ -3,8 +3,8 @@ class VariableDec {
     int a=0;
     int b=0;
     int c=a+b;
-    float d=1;
-    float e=c+d;
+    double d=1;
+    double e=c+d;
 
   }
 }
